loggen/requirements.txt
opentelemetry.api
opentelemetry.sdk
opentelemetry.exporter.otlp.proto.grpc
opentelemetry.instrumentation.logging
